Implementando Teste

+++ planeapp/models/produto.py
@@ -11,7 +11,7 @@ class Produto(models.Model):
     '''
 
     nome = models.CharField(max_length=100)
-    preco_unit = models.FloatField()
+    preco_unit = models.DecimalField(max_digits=10,decimal_places=2)
     multiplo = models.PositiveIntegerField(default=1,blank=True)
     imagem = models.ImageField(upload_to="produto",blank=True)
